+++ abci/app/v1/check_tx.go
@@ -95,12 +95,14 @@ var IsMethod = map[string]bool{
 	"AddIdentity":                      true,
 	"SetAllowedModeList":               true,
 	"UpdateNamespace":                  true,
-	"SetAllowedMinIalForRegisterIdentityAtFirstIdp": true,
-	"RevokeAndAddAccessor":                          true,
-	"SetServicePriceCeiling":                        true,
-	"SetServicePriceMinEffectiveDatetimeDelay":      true,
-	"SetServicePrice":                               true,
-	"CreateMessage":                                 true,
+	"SetAllowedMinIalForRegisterIdentityAtFirstIdp":        true,
+	"RevokeAndAddAccessor":                                 true,
+	"SetServicePriceCeiling":                               true,
+	"SetServicePriceMinEffectiveDatetimeDelay":             true,
+	"SetServicePrice":                                      true,
+	"CreateMessage":                                        true,
+	"AddSuppressedIdentityModificationNotificationNode":    true,
+	"RemoveSuppressedIdentityModificationNotificationNode": true,
 }
 
 func (app *ABCIApplication) checkTxInitNDID(param string, nodeID string) types.ResponseCheckTx {
@@ -713,7 +715,9 @@ func (app *ABCIApplication) callCheckTx(name string, param string, nodeID string
 		"UpdateNamespace",
 		"SetAllowedMinIalForRegisterIdentityAtFirstIdp",
 		"SetServicePriceCeiling",
-		"SetServicePriceMinEffectiveDatetimeDelay":
+		"SetServicePriceMinEffectiveDatetimeDelay",
+		"AddSuppressedIdentityModificationNotificationNode",
+		"RemoveSuppressedIdentityModificationNotificationNode":
 		return app.checkIsNDID(param, nodeID)
 	case "RegisterIdentity",
 		"AddAccessor",