+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## TBD
+
+FEATURES:
+
+- Suppressed identity modification notification node list
+  - Add `AddSuppressedIdentityModificationNotificationNode` method.
+  - Add `RemoveSuppressedIdentityModificationNotificationNode` method.
+  - [Query] Add `GetSuppressedIdentityModificationNotificationNodeList` method.
+  - [Query] Add `IsSuppressedIdentityModificationNotificationNode` method.
+
 ## 6.0.0 (July 27, 2021)
 
 IMPROVEMENTS: